fix(control): remove the race condition within the (*Query).Done method (#1277)
If cancel and done were called in quick succession, it could result in a
race condition because there was no lock over the state transition
inside of done. It was assumed that done was called after results were
read, but that's not part of the interface and you are allowed to call
done without having read the results. This became most obvious when
cancel was called and then done immediately afterwards.

This modifies it so that cancel is asynchronous and signals that the
execution should be halted, which is how it was previously. It modifies
done to wait for the task to actually be completed before the call
returns.