fix(stdlib/map): remove mergeKey parameter from the map function
BREAKING CHANGE: The removal of the mergeKey parameter is a breaking
change. Users will either need to use the `with` operator to preserve
columns or explicitly list all columns they wish to preserve through the
map function.
Several other test cases that used map have been simplified to remove
their use of the map function as we want the test cases to focus on
testing the function they are designed to test.