feat: add enriched error interface to flux (#1423)
We need a common way of categorizing and understanding errors returned
from InfluxDB. We model the errors based on the same concepts that
InfluxDB uses, but we create our own error codes that are modeled on the
grpc error codes.

Each native flux error will have a code that represents the category or
type of error and a message that gives detailed and human-readable
output about the error.

The errors library is included as an internal library and the error
struct itself is exposed using a type alias in the top level package.
This allows the errors to be used by packages that cannot depend on the
top level package (like semantic), but still have the error type be
exposed.